refactor serving gui
