+++ supervisely/nn/inference/gui/__init__.py
@@ -1,1 +1,2 @@
-from supervisely.nn.inference.gui.gui import BaseInferenceGUI, InferenceGUI, ServingGUI
+from supervisely.nn.inference.gui.gui import BaseInferenceGUI, InferenceGUI
+from supervisely.nn.inference.gui.serving_gui import ServingGUI
